Builtin and Specification
* fixed a compilation issue regarding the built-ins for older gcc/clang
  versions
* fixed inheritance of Specification class

+++ src/Builtin.cpp
@@ -40,7 +40,7 @@ Builtin::Builtin(
 {
     getSymbols()[ ".builtin" ].insert( this );
 
-    id2obj()[ id ] = this;
+    id2obj()[ (u8)id ] = this;
     str2obj()[ std::string( name ) ] = this;
 }
